Move PyPackageUtil into sepearet file
